# Clean-up / simplify plotting code