--- a/geopandas/plotting.py
+++ b/geopandas/plotting.py
@@ -47,9 +47,9 @@ def _flatten_multi_geoms(geoms, colors):
     return components, component_colors
 
 
-def plot_polygon_collection(ax, geoms, colors_or_values, plot_values,
-                            vmin=None, vmax=None, cmap=None,
-                            edgecolor='black', alpha=0.5, linewidth=1.0, **kwargs):
+def plot_polygon_collection(ax, geoms, values=None, linewidth=1.0,
+                            edgecolor='black', alpha=0.5,
+                            vmin=None, vmax=None, cmap=None, **kwargs):
     """
     Plots a collection of Polygon and MultiPolygon geometries to `ax`
 
@@ -61,51 +61,45 @@ def plot_polygon_collection(ax, geoms, colors_or_values, plot_values,
 
     geoms : a sequence of `N` Polygons and/or MultiPolygons (can be mixed)
 
-    colors_or_values : a sequence of `N` values or RGBA tuples
-        It should have 1:1 correspondence with the geometries (not their components).
+    values : a sequence of `N` values
+        Values will be mapped to colors using vmin/vmax/cmap. They should
+        have 1:1 correspondence with the geometries (not their components).
+        Otherwise follows `color` / `facecolor` kwargs.
 
-    plot_values : bool
-        If True, `colors_or_values` is interpreted as a list of values, and will
-        be mapped to colors using vmin/vmax/cmap (which become required).
-        Otherwise `colors_or_values` is interpreted as a list of colors.
+    edgecolor : single color or sequence of `N` colors
+        Color for the edge of the polygons
+
+    facecolor : single color or sequence of `N` colors
+        Color to fill the polygons. Cannot be used together with `values`.
+
+    color : single color or sequence of `N` colors
+        Sets both `edgecolor` and `facecolor`
+
+    **kwargs
+        Additional keyword arguments passed to the collection
 
     Returns
     -------
 
     collection : matplotlib.collections.Collection that was plotted
     """
-
     from descartes.patch import PolygonPatch
     from matplotlib.collections import PatchCollection
 
-    components, component_colors_or_values = _flatten_multi_geoms(
-        geoms, colors_or_values)
+    geoms, values = _flatten_multi_geoms(geoms, values)
 
     # PatchCollection does not accept some kwargs.
     if 'markersize' in kwargs:
         del kwargs['markersize']
-    collection = PatchCollection([PolygonPatch(poly) for poly in components],
+
+    collection = PatchCollection([PolygonPatch(poly) for poly in geoms],
                                  linewidth=linewidth, edgecolor=edgecolor,
                                  alpha=alpha, **kwargs)
 
-    if plot_values:
-        collection.set_array(np.array(component_colors_or_values))
+    if values is not None:
+        collection.set_array(np.asarray(values))
         collection.set_cmap(cmap)
         collection.set_clim(vmin, vmax)
-    else:
-        # set_color magically sets the correct combination of facecolor and
-        # edgecolor, based on collection type.
-        collection.set_color(component_colors_or_values)
-
-        # If the user set facecolor and/or edgecolor explicitly, the previous
-        # call to set_color might have overridden it (remember, the 'color' may
-        # have come from plot_series, not from the user). The user should be
-        # able to override matplotlib's default behavior, by setting them again
-        # after set_color.
-        if 'facecolor' in kwargs:
-            collection.set_facecolor(kwargs['facecolor'])
-        if edgecolor:
-            collection.set_edgecolor(edgecolor)
 
     ax.add_collection(collection, autolim=True)
     ax.autoscale_view()
@@ -141,8 +135,7 @@ def plot_linestring_collection(ax, geoms, values=None, color=None,
     """
     from matplotlib.collections import LineCollection
 
-    geoms, values = _flatten_multi_geoms(
-        geoms, values)
+    geoms, values = _flatten_multi_geoms(geoms, values)
 
     # LineCollection does not accept some kwargs.
     if 'markersize' in kwargs: