I went through the plotting code and did some clean-up (builds further upon the work of @IamJeffG):

- Changed the signature of the `plot_*_collection` functions. They now just have an optional `values=None` argument (instead of `colors_or_values, plot_values` combo where `plot_values` indicated whether `color_or_values` should be interpreted as colors or as values
  - This makes the implementation IMO a bit cleaner
  - This simplifies the code as I leave all handling of the color kwargs to matplotlib collection constructors (no need to set them manually)
- I removed the legacy handling of `alpha` for plotting polygons:
  - Currently polygons were plotted twice to be able to apply alpha to fill color, but not to edge. Removed this hack, which implies of course a behavioural change that the alpha is now applied on both edge and fill.
  - This cleaned up the code a lot, and also improves the performance for plotting polygons
- Some other clean-up, eg not re-computing the `geometry.types` will improve performance a little bit (will also do a PEP8)

Result from the benchmarks also indicate that their is some speed improvement, mainly for polygons:

```
       before           after         ratio
     [baa1601c]       [a951e68c]
-        73.1±2ms         64.9±1ms     0.89  plotting.Bench.time_plot_values('mixed')
-        99.4±3ms         87.8±3ms     0.88  plotting.Bench.time_plot_values('Point')
-        71.4±3ms         62.2±2ms     0.87  plotting.Bench.time_plot_series('LineString')
-         105±1ms         87.8±1ms     0.83  plotting.Bench.time_plot_series('mixed')
-        82.5±3ms       62.9±0.8ms     0.76  plotting.Bench.time_plot_values('MultiPolygon')
-         104±2ms         73.4±1ms     0.71  plotting.Bench.time_plot_values('Polygon')
-        174±30ms         96.7±8ms     0.56  plotting.Bench.time_plot_series('Polygon')
-         122±2ms         65.3±1ms     0.54  plotting.Bench.time_plot_series('MultiPolygon')
```

Those changes should mainly be back compatible (apart from the alpha change, the only other behavioural change (based on the tests) is that now color overrules facecolor instead of the other way around for polygons (this follows matplotlib behaviour). 
Further, there is one test that is not yet passing for MultiPolygons. I can fix this one, but if we decide in https://github.com/geopandas/geopandas/issues/318 to switch to uniform colors by default, this fix is not needed (so will wait on that)

But of course, not everything will probably be covered in the tests, so feedback / testing welcome!

cc @IamJeffG @jdmcbr 

(it might be easier to digest if you look at the diff commit by commit, there is also some explanation in the commit message) 
And sorry for doing this so close to the release (want to release next week)

cc @darribas @emiliom always welcome to run some tutorial material with this branch to see if there are problems.