@jorisvandenbossche 2017-08-23
I am going to merge this, so I can do a new PR with the style changes (I want to do that in a separate PR, as there I will need to change more tests). 

@jdmcbr (or anyone else) if you still have comments, feel free to leave them here, and then I can include the fixes in the new PR.