@jdmcbr 2017-08-24
I think with #510, this is a big step towards easier to use and maintain plotting tools.