@darribas 2017-08-25
This looks amazing!!! Unfortunately, too close to the start of my course to get it on [GDS](http://darribas.org/gds17) this year as the university system won't let me update it. But it'll definitely go on the next iteration!